requirements.txt
django>=5.0,<7.0
django-allauth>=65.0
djangorestframework>=3.15
qrcode[pil]>=8.0
user-agents>=2.2
python-dotenv>=1.0
whitenoise>=6.5
gunicorn>=21.2
PyJWT>=2.8
cryptography>=41.0
